added a global file but got a linear gradient problem

+++ lib/workout.dart
@@ -2,8 +2,8 @@ class Workout {
   String? exerciseName;
   int? reps;
   int? sets;
-  int? weight;
-  int? rest;
+  double? weight;
+  double? rest;
 
   Workout(this.reps, this.sets, this.weight, this.exerciseName, this.rest);
 }
